+++ docs.json
@@ -1,0 +1,10 @@
+{
+    "name": "docs.page",
+    "theme": "#36B9B9",
+    "sidebar": [
+        [
+            "Home",
+            "/"
+        ]
+    ]
+}
