+++ docs/index.mdx
@@ -1,0 +1,2 @@
+# Example Docs!
+
