+++ lib/main_development.dart
@@ -1,0 +1,3 @@
+import 'package:mobile_teset/app.dart';
+
+void main() => entry('development');
