+++ lib/main_production.dart
@@ -1,0 +1,3 @@
+import 'package:mobile_teset/app.dart';
+
+void main() => entry('production');
